Agent/requirements.txt
# LangGraph SQL Agent Dependencies
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# MCP Integration
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# 可视化 (本地图表生成)
pyecharts>=2.0.0
rich>=13.0.0

